service: complete report service and merge stats to report repo
