+++ services/api/internal/services/ocserv_user/types.go
@@ -35,21 +35,6 @@ type OcservUsersResponse struct {
 	Result []models.OcservUser `json:"result" validate:"omitempty"`
 }
 
-type StatisticsData struct {
-	DateStart string `json:"date_start" query:"date_start" validate:"omitempty" example:"2025-1-31"`
-	DateEnd   string `json:"date_end" query:"date_end" validate:"omitempty" example:"2025-12-31"`
-}
-
-type StatisticsResponse struct {
-	Statistics      []models.DailyTraffic      `json:"statistics" validate:"required"`
-	TotalBandwidths repository.TotalBandwidths `json:"total_bandwidths" validate:"required"`
-}
-
-type TotalBandwidthData struct {
-	DateStart string `json:"date_start" query:"date_start" validate:"omitempty" example:"2025-1-31"`
-	DateEnd   string `json:"date_end" query:"date_end" validate:"omitempty" example:"2025-12-31"`
-}
-
 type SyncOcpasswdRequest struct {
 	Users       []user.Ocpasswd          `json:"users" validate:"required"`
 	ExpireAt    *string                  `json:"expire_at" validate:"omitempty" example:"2025-12-31"`
@@ -68,13 +53,6 @@ type ActivateUserData struct {
 	ExpireAt *string `json:"expire_at" validate:"omitempty" example:"2025-12-31"`
 }
 
-type UserStatsResponse struct {
-	Online      int   `json:"online"`
-	Active      int64 `json:"active"`
-	Deactivated int64 `json:"deactivated"`
-	Locked      int64 `json:"locked"`
-}
-
 type SessionLogsData struct {
 	DateStart string `json:"date_start" query:"date_start" validate:"omitempty" example:"2025-1-31"`
 	DateEnd   string `json:"date_end" query:"date_end" validate:"omitempty" example:"2025-12-31"`
@@ -83,4 +61,14 @@ type SessionLogsData struct {
 type SessionLogsResponse struct {
 	Meta   request.Meta                   `json:"meta" validate:"required"`
 	Result *[]models.OcservUserSessionLog `json:"result" validate:"omitempty"`
+}
+
+type StatisticsData struct {
+	DateStart string `json:"date_start" query:"date_start" validate:"omitempty" example:"2025-1-31"`
+	DateEnd   string `json:"date_end" query:"date_end" validate:"omitempty" example:"2025-12-31"`
+}
+
+type StatisticsResponse struct {
+	Statistics      []models.DailyTraffic      `json:"statistics" validate:"required"`
+	TotalBandwidths repository.TotalBandwidths `json:"total_bandwidths" validate:"required"`
 }
