+++ services/api/internal/services/report/controller.go
@@ -1,23 +1,28 @@
-package statistics
+package report
 
 import (
+	"errors"
 	"fmt"
 	"github.com/labstack/echo/v4"
 	"github.com/mmtaee/ocserv-dashboard/api/internal/repository"
 	"github.com/mmtaee/ocserv-dashboard/api/pkg/request"
 	"net/http"
+	"strings"
+	"sync"
 	"time"
 )
 
 type Controller struct {
-	request    request.CustomRequestInterface
-	reportRepo repository.ReportRepositoryInterface
+	request         request.CustomRequestInterface
+	reportRepo      repository.ReportRepositoryInterface
+	ocservOcctlRepo repository.OcctlRepositoryInterface
 }
 
 func New() *Controller {
 	return &Controller{
-		request:    request.NewCustomRequest(),
-		reportRepo: repository.NewtReportRepository(),
+		request:         request.NewCustomRequest(),
+		reportRepo:      repository.NewtReportRepository(),
+		ocservOcctlRepo: repository.NewOcctlRepository(),
 	}
 }
 
@@ -78,5 +83,167 @@ func (ctl *Controller) SessionLogs(c echo.Context) error {
 			PageSize:     pagination.PageSize,
 		},
 		Result: logs,
+	})
+}
+
+// Statistics 	 Ocserv Users Statistics
+//
+// @Summary      Ocserv Users Statistics
+// @Description  Ocserv Users Statistics
+// @Tags         Report
+// @Accept       json
+// @Produce      json
+// @Param        Authorization header string true "Bearer TOKEN"
+// @Param 		 date_start query string true "date_start"
+// @Param 		 date_end query string true "date_end"
+// @Failure      400 {object} request.ErrorResponse
+// @Failure      401 {object} middlewares.Unauthorized
+// @Success      200 {object} []models.DailyTraffic
+// @Router       /reports/statistics [get]
+func (ctl *Controller) Statistics(c echo.Context) error {
+	var data StatisticsData
+	if err := c.Bind(&data); err != nil {
+		return ctl.request.BadRequest(c, err)
+	}
+
+	if data.DateStart == "" || data.DateEnd == "" {
+		return ctl.request.BadRequest(c, errors.New("statistics date start and end are required"))
+	}
+
+	var startDate, endDate *time.Time
+
+	tStart, err := time.Parse("2006-01-02", data.DateStart)
+	if err != nil {
+		return ctl.request.BadRequest(c, fmt.Errorf("invalid date_start: %w", err))
+	}
+	startDate = &tStart
+
+	tEnd, err := time.Parse("2006-01-02", data.DateEnd)
+	if err != nil {
+		return ctl.request.BadRequest(c, fmt.Errorf("invalid date_end: %w", err))
+	}
+	tEnd = tEnd.Add(23*time.Hour + 59*time.Minute + 59*time.Second + 999999999*time.Nanosecond)
+	endDate = &tEnd
+
+	if tStart.After(*endDate) {
+		return ctl.request.BadRequest(c, errors.New("date start is after end"))
+	}
+
+	stats, err := ctl.reportRepo.Statistics(c.Request().Context(), startDate, endDate)
+	if err != nil {
+		return ctl.request.BadRequest(c, err)
+	}
+	return c.JSON(http.StatusOK, stats)
+}
+
+// TotalBandwidth 	 Ocserv Users TotalBandwidth calculating
+//
+// @Summary      Ocserv Users TotalBandwidth calculating
+// @Description  Ocserv Users TotalBandwidth calculating
+// @Tags         Report
+// @Accept       json
+// @Produce      json
+// @Param        Authorization header string true "Bearer TOKEN"
+// @Param 		 date_start query string true "date_start"
+// @Param 		 date_end query string true "date_end"
+// @Failure      400 {object} request.ErrorResponse
+// @Failure      401 {object} middlewares.Unauthorized
+// @Success      200 {object} repository.TotalBandwidths
+// @Router       /reports/total-bandwidth [get]
+func (ctl *Controller) TotalBandwidth(c echo.Context) error {
+	var data TotalBandwidthData
+	if err := c.Bind(&data); err != nil {
+		return ctl.request.BadRequest(c, err)
+	}
+
+	var startDate, endDate *time.Time
+
+	if data.DateStart != "" {
+		t, err := time.Parse("2006-01-02", data.DateStart)
+		if err != nil {
+			return ctl.request.BadRequest(c, fmt.Errorf("invalid date_start: %w", err))
+		}
+		startDate = &t
+	}
+
+	if data.DateEnd != "" {
+		t, err := time.Parse("2006-01-02", data.DateEnd)
+		if err != nil {
+			return ctl.request.BadRequest(c, fmt.Errorf("invalid date_end: %w", err))
+		}
+		t = t.Add(23*time.Hour + 59*time.Minute + 59*time.Second + 999999999*time.Nanosecond)
+		endDate = &t
+	}
+
+	if startDate != nil && endDate != nil && startDate.After(*endDate) {
+		return ctl.request.BadRequest(c, errors.New("date start is after end"))
+	}
+
+	bandwidth, err := ctl.reportRepo.TotalBandwidthDateRange(c.Request().Context(), startDate, endDate)
+	if err != nil {
+		return ctl.request.BadRequest(c, err)
+	}
+	return c.JSON(http.StatusOK, bandwidth)
+}
+
+// OcservUserReport     Result of all user reports
+//
+// @Summary      Result of all user reports
+// @Description  Result of all user reports
+// @Tags         Report
+// @Accept       json
+// @Produce      json
+// @Param        Authorization header string true "Bearer TOKEN"
+// @Failure      400 {object} request.ErrorResponse
+// @Failure      401 {object} middlewares.Unauthorized
+// @Success      200 {object} OcservUserReportResponse
+// @Router       /reports/users [get]
+func (ctl *Controller) OcservUserReport(c echo.Context) error {
+	var wg sync.WaitGroup
+	var onlineUsers []string
+	var result repository.UserStatsResult
+
+	errChan := make(chan error, 2)
+	wg.Add(2)
+
+	go func() {
+		defer wg.Done()
+
+		users, err := ctl.ocservOcctlRepo.OnlineUsers()
+		if err != nil {
+			errChan <- fmt.Errorf("failed to get online users: %w", err)
+			return
+		}
+		onlineUsers = users
+	}()
+
+	go func() {
+		defer wg.Done()
+
+		res, err := ctl.reportRepo.UsersStat(c.Request().Context())
+		if err != nil {
+			errChan <- fmt.Errorf("failed to get users stats: %w", err)
+			return
+		}
+		result = res
+	}()
+
+	wg.Wait()
+	close(errChan)
+
+	var errs []string
+	for e := range errChan {
+		errs = append(errs, e.Error())
+	}
+
+	if len(errs) > 0 {
+		return ctl.request.BadRequest(c, errors.New(strings.Join(errs, "; ")))
+	}
+
+	return c.JSON(http.StatusOK, OcservUserReportResponse{
+		Online:      len(onlineUsers),
+		Active:      result.Active,
+		Deactivated: result.Deactivated,
+		Locked:      result.Locked,
 	})
 }
