+++ services/api/internal/services/report/routes.go
@@ -1,4 +1,4 @@
-package statistics
+package report
 
 import (
 	"github.com/labstack/echo/v4"
@@ -7,7 +7,10 @@ import (
 
 func Routes(e *echo.Group) {
 	ctl := New()
-	g := e.Group("/reports", middlewares.AuthMiddleware())
+	g := e.Group("/reports", middlewares.AuthMiddleware(), middlewares.AdminPermission())
 
 	g.GET("/session_logs", ctl.SessionLogs)
+	g.GET("/statistics", ctl.Statistics)
+	g.GET("/users", ctl.OcservUserReport)
+	g.GET("/total-bandwidth", ctl.TotalBandwidth)
 }
